# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Articles API",
    "description": "A RESTful API for managing articles and authors, built with NestJS following Clean Architecture principles and CQRS pattern. Features include article management, author management, and JWT-based authentication with OTP verification.",
    "version": "1.0.0",
    "contact": {
      "name": "Iliyas Akbergen",
      "url": "https://github.com/IlyasAkbergen",
      "email": "iliyas.akbergen@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3030",
      "description": "Local development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/articles": {
      "get": {
        "tags": ["Articles"],
        "summary": "Get all articles with pagination",
        "description": "Retrieves a paginated list of articles with optional filtering and sorting.",
        "operationId": "getAllArticles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["createdAt", "updatedAt", "title", "publishedAt"],
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            },
            "description": "Sort order"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search term for title or content"
          },
          {
            "name": "authorId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by author ID"
          },
          {
            "name": "published",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by publication status"
          }
        ],
        "responses": {
          "200": {
            "description": "Articles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedArticlesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": ["Articles"],
        "summary": "Create a new article",
        "description": "Creates a new article with the provided title, content, and author ID.",
        "operationId": "createArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Article created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Get article by ID",
        "description": "Retrieves a specific article by its unique identifier.",
        "operationId": "getArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Article unique identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Article retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "put": {
        "tags": ["Articles"],
        "summary": "Update an article",
        "description": "Updates an existing article with new title and/or content.",
        "operationId": "updateArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Article unique identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": ["Articles"],
        "summary": "Delete an article",
        "description": "Permanently deletes an article from the system.",
        "operationId": "deleteArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Article unique identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Article deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/articles/{id}/publish": {
      "post": {
        "tags": ["Articles"],
        "summary": "Publish an article",
        "description": "Changes the status of an article to published, making it visible to readers.",
        "operationId": "publishArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Article unique identifier"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article published successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "400": {
            "description": "Article already published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/articles/{id}/unpublish": {
      "post": {
        "tags": ["Articles"],
        "summary": "Unpublish an article",
        "description": "Changes the status of an article to unpublished, hiding it from readers.",
        "operationId": "unpublishArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Article unique identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Article unpublished successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "400": {
            "description": "Article already unpublished",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/articles/author/{authorId}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Get articles by author",
        "description": "Retrieves all articles written by a specific author.",
        "operationId": "getArticlesByAuthor",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Author unique identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Articles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResponse"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "CreateArticleDto": {
        "type": "object",
        "required": ["title", "content", "authorId"],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "description": "The title of the article",
            "example": "Understanding Clean Architecture in NestJS"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "description": "The content of the article in markdown format",
            "example": "This article explores the principles of Clean Architecture and how to implement them in a NestJS application..."
          },
          "authorId": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier of the author",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        }
      },
      "UpdateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "description": "The updated title of the article",
            "example": "Understanding Clean Architecture in NestJS - Updated"
          },
          "content": {
            "type": "string",
            "description": "The updated content of the article in markdown format",
            "example": "This updated article explores the principles of Clean Architecture..."
          }
        }
      },
      "PublishArticleDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Optional message for publishing the article",
            "example": "Article is ready for publication"
          }
        }
      },
...
      "ArticleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the article",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "description": "Title of the article",
            "example": "Understanding Clean Architecture in NestJS"
          },
          "content": {
            "type": "string",
            "description": "Content of the article in markdown format",
            "example": "This article explores the principles of Clean Architecture and how to implement them in a NestJS application..."
          },
          "authorId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the author",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "authorName": {
            "type": "string",
            "description": "Full name of the author",
            "example": "Iliyas Akbergen"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Whether the article is published",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the article was created",
            "example": "2025-08-08T10:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the article was last updated",
            "example": "2025-08-08T12:00:00Z"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date when the article was published",
            "example": "2025-08-08T12:00:00Z"
          },
          "wordCount": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of words in the article content",
            "example": 150
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1,
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "description": "Number of items per page",
            "example": 10
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "description": "Total number of items",
            "example": 25
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "description": "Total number of pages",
            "example": 3
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether there is a next page",
            "example": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Whether there is a previous page",
            "example": false
          }
        }
      },
      "PaginatedArticlesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleResponse"
            },
            "description": "Array of paginated articles"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Field is required", "Invalid format"]
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - validation errors",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - invalid or missing JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}